Create dynamic time for job post (#22)
What is this PR about?
Previously, our time ago is not dynamic, and this can actually cause confusion.
Imagine posting a job that was created 30 mins ago to discord. And then 2 days later, you scroll back and it still shows This job was posted 30 minutes ago.
This PR attempts to fix this by using Discord's timestamp.